+++ cirq-core/cirq/sim/simulator_base_test.py
@@ -12,10 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import math
-from typing import List, Dict, Any, Sequence, Tuple, Union
+from typing import List, Dict, Any, Sequence, Union
 
 import numpy as np
 import pytest
+import sympy
 
 import cirq
 
@@ -356,3 +357,46 @@ def test_measurements_retained_in_step_results():
     assert next(iterator).measurements.keys() == {'a', 'b'}
     assert next(iterator).measurements.keys() == {'a', 'b', 'c'}
     assert not any(iterator)
+
+
+def test_sweep_unparameterized_prefix_not_repeated_iff_unitary():
+    q = cirq.LineQubit(0)
+
+    class TestOp(cirq.Operation):
+        def __init__(self, *, has_unitary: bool):
+            self.count = 0
+            self.has_unitary = has_unitary
+
+        def _act_on_(self, args):
+            self.count += 1
+            return True
+
+        def with_qubits(self, qubits):
+            pass
+
+        @property
+        def qubits(self):
+            return (q,)
+
+        def _has_unitary_(self):
+            return self.has_unitary
+
+    simulator = CountingSimulator()
+    params = [
+        cirq.ParamResolver({'a': 0}),
+        cirq.ParamResolver({'a': 1}),
+    ]
+
+    op1 = TestOp(has_unitary=True)
+    op2 = TestOp(has_unitary=True)
+    circuit = cirq.Circuit(op1, cirq.XPowGate(exponent=sympy.Symbol('a'))(q), op2)
+    simulator.simulate_sweep(program=circuit, params=params)
+    assert op1.count == 1
+    assert op2.count == 2
+
+    op1 = TestOp(has_unitary=False)
+    op2 = TestOp(has_unitary=False)
+    circuit = cirq.Circuit(op1, cirq.XPowGate(exponent=sympy.Symbol('a'))(q), op2)
+    simulator.simulate_sweep(program=circuit, params=params)
+    assert op1.count == 2
+    assert op2.count == 2
