+++ cirq-core/cirq/sim/sparse_simulator_test.py
@@ -1332,36 +1332,3 @@ def test_noise_model():
     result = simulator.run(circuit, repetitions=100)
 
     assert 20 <= sum(result.measurements['0'])[0] < 80
-
-
-def test_simulate_sweep_parameters_prefix():
-    q = cirq.LineQubit(0)
-
-    class TestOp(cirq.Operation):
-        count = 0
-
-        def _act_on_(self, args):
-            self.count += 1
-            return True
-
-        def with_qubits(self, qubits):
-            pass
-
-        @property
-        def qubits(self):
-            return (q,)
-
-        def _has_unitary_(self):
-            return True
-
-    simulator = cirq.Simulator()
-    op1 = TestOp()
-    op2 = TestOp()
-    circuit = cirq.Circuit(op1, cirq.XPowGate(exponent=sympy.Symbol('a'))(q), op2)  # type: ignore
-    params = [
-        cirq.ParamResolver({'a': 0}),
-        cirq.ParamResolver({'a': 1}),
-    ]
-    simulator.simulate_sweep(program=circuit, params=params)
-    assert op1.count == 1
-    assert op2.count == 2
